+++ TinyBoss/perceptron.h
@@ -18,19 +18,19 @@ using namespace std;
 class classPerceptron {
 private:
     int ni;
-    vector <float> ponderation;
-    vector <float> inputs;
-    float output;
-    float mu = 0.6;
-    float dEtotaldOutput = 0.0;
+    vector <double> ponderation;
+    vector <double> inputs;
+    double output;
+    double mu = 0.6;
+    double dEtotaldOutput = 0.0;
     
 public:
-    classPerceptron(int nip, float initialMu);
+    classPerceptron(int nip, double initialMu);
     ~classPerceptron();
-    int setInputs(vector <float> input);
-    float getOutput();
-    void setError(float error);
-    vector<float> fix(vector<float> inputs);
+    int setInputs(vector <double> input);
+    double getOutput();
+    void setError(double error);
+    vector<double> fix(vector<double> inputs);
 };
 
 #endif /* perceptron_h */
